refactor(taskflow): extract feature.collaborators
Move CollaboratorsModel, collaboratorsReducer, seedCollaborators, and
BotCollaborator (startBot) into …feature.collaborators; update all
consumers (AccountStore, AccountRegistry, BoardScreen, CardDetailScreen,
ProfileScreen, AccountReducersTest) to import from the new package.
