+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/model/AccountModels.kt
@@ -4,19 +4,16 @@ import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.PersistentMap
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentMapOf
-import org.reduxkotlin.sample.taskflow.core.AccountId
-import org.reduxkotlin.sample.taskflow.core.AccountSummary
 import org.reduxkotlin.sample.taskflow.core.BoardId
 import org.reduxkotlin.sample.taskflow.core.BoardSummary
 
 // --- Per-account models (ALL declared up front; board slices start empty/sentinel) ---
 // Identity (name/email/avatar) is NOT duplicated here — it lives once in CollaboratorsModel
 // (which includes self). SessionModel moved to …feature.account.AccountModels.
+// CollaboratorsModel moved to …feature.collaborators.CollaboratorsModels.
 
-data class BoardListModel(
+/** Holds the cached board tiles and their display order for one account. */
+public data class BoardListModel(
     val boards: PersistentMap<BoardId, BoardSummary> = persistentMapOf(),
     val order: PersistentList<BoardId> = persistentListOf(),
 )
-
-// Resolves assignee/creator/bot avatars within the account store (no root reach-in).
-data class CollaboratorsModel(val byId: PersistentMap<AccountId, AccountSummary> = persistentMapOf())
